# store expiring addrs by pointer