Hopefully, this reduces the space wasted by using maps (see the comment in the committed code).