@Stebalien 2018-10-15
@rob-deutsch, could you try testing this patch? Hypothesis: unused slots in `map[string]*expiringAddr` should be smaller (one ptr versus 6) than `map[string]expiringAddr`. This *will* increase the *number* of allocations quite a bit but should hopefully reduce allocations overall (unless the heap overhead/fragmentation of having more objects kills the performance gains).