@raulk 2018-10-15
@Stebalien we have benchmarks now, in case those are useful useful to run against this diff:

https://github.com/libp2p/go-libp2p-peerstore/blob/1b410c7294a896fe5a1077d6b049178413ab6d9d/pstoremem/inmem_test.go#L28