@Kubuxu 2018-10-15
I've run benchmarks:
```
benchmark                                                             old ns/op     new ns/op     delta
BenchmarkInMemoryPeerstore/AddAddrs-1Addrs-InMem-4                    21806         21732         -0.34%
BenchmarkInMemoryPeerstore/AddAddrs-10Addrs-InMem-4                   64750         65084         +0.52%
BenchmarkInMemoryPeerstore/AddAddrs-100Addrs-InMem-4                  506152        492290        -2.74%
BenchmarkInMemoryPeerstore/SetAddrs-1Addrs-InMem-4                    20975         21981         +4.80%
BenchmarkInMemoryPeerstore/SetAddrs-10Addrs-InMem-4                   63849         65278         +2.24%
BenchmarkInMemoryPeerstore/SetAddrs-100Addrs-InMem-4                  488247        498307        +2.06%
BenchmarkInMemoryPeerstore/GetAddrs-1Addrs-InMem-4                    262           256           -2.29%
BenchmarkInMemoryPeerstore/GetAddrs-10Addrs-InMem-4                   677           696           +2.81%
BenchmarkInMemoryPeerstore/GetAddrs-100Addrs-InMem-4                  4692          4813          +2.58%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-1Addrs-InMem-4         22694         23609         +4.03%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-10Addrs-InMem-4        67677         70709         +4.48%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-100Addrs-InMem-4       506751        476841        -5.90%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-1Addrs-InMem-4       136159        100476        -26.21%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-10Addrs-InMem-4      117103        100919        -13.82%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-100Addrs-InMem-4     118497        101303        -14.51%

benchmark                                                             old allocs     new allocs     delta
BenchmarkInMemoryPeerstore/AddAddrs-1Addrs-InMem-4                    35             36             +2.86%
BenchmarkInMemoryPeerstore/AddAddrs-10Addrs-InMem-4                   240            250            +4.17%
BenchmarkInMemoryPeerstore/AddAddrs-100Addrs-InMem-4                  2312           2411           +4.28%
BenchmarkInMemoryPeerstore/SetAddrs-1Addrs-InMem-4                    35             36             +2.86%
BenchmarkInMemoryPeerstore/SetAddrs-10Addrs-InMem-4                   240            250            +4.17%
BenchmarkInMemoryPeerstore/SetAddrs-100Addrs-InMem-4                  2312           2410           +4.24%
BenchmarkInMemoryPeerstore/GetAddrs-1Addrs-InMem-4                    1              1              +0.00%
BenchmarkInMemoryPeerstore/GetAddrs-10Addrs-InMem-4                   1              1              +0.00%
BenchmarkInMemoryPeerstore/GetAddrs-100Addrs-InMem-4                  1              1              +0.00%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-1Addrs-InMem-4         36             37             +2.78%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-10Addrs-InMem-4        242            252            +4.13%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-100Addrs-InMem-4       2315           2411           +4.15%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-1Addrs-InMem-4       1              1              +0.00%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-10Addrs-InMem-4      2              3              +50.00%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-100Addrs-InMem-4     22             22             +0.00%

benchmark                                                             old bytes     new bytes     delta
BenchmarkInMemoryPeerstore/AddAddrs-1Addrs-InMem-4                    7749          7426          -4.17%
BenchmarkInMemoryPeerstore/AddAddrs-10Addrs-InMem-4                   18913         18649         -1.40%
BenchmarkInMemoryPeerstore/AddAddrs-100Addrs-InMem-4                  134942        133411        -1.13%
BenchmarkInMemoryPeerstore/SetAddrs-1Addrs-InMem-4                    7746          7426          -4.13%
BenchmarkInMemoryPeerstore/SetAddrs-10Addrs-InMem-4                   18914         18650         -1.40%
BenchmarkInMemoryPeerstore/SetAddrs-100Addrs-InMem-4                  134947        133383        -1.16%
BenchmarkInMemoryPeerstore/GetAddrs-1Addrs-InMem-4                    16            16            +0.00%
BenchmarkInMemoryPeerstore/GetAddrs-10Addrs-InMem-4                   160           160           +0.00%
BenchmarkInMemoryPeerstore/GetAddrs-100Addrs-InMem-4                  1792          1792          +0.00%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-1Addrs-InMem-4         7625          7306          -4.18%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-10Addrs-InMem-4        19011         18750         -1.37%
BenchmarkInMemoryPeerstore/AddGetAndClearAddrs-100Addrs-InMem-4       136802        135078        -1.26%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-1Addrs-InMem-4       65569         49219         -24.94%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-10Addrs-InMem-4      65619         49318         -24.84%
BenchmarkInMemoryPeerstore/Get1000PeersWithAddrs-100Addrs-InMem-4     50308         50308         +0.00%
```