@raulk 2018-10-15
Team: Let's hold off on merging until @Stebalien has a chance to review the benchmark diffs kindly posted by @Kubuxu!