@Stebalien 2018-10-16
The benchmarks look good to me. Allocations are slightly up but memory usage is slightly down. IMO, this is worth merging.